email ai agent tutorial/requirements.txt
langchain
langchain-ollama
langgraph
python-dotenv
imap-client
